requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask[async]==2.3.3
gunicorn==21.2.0

# Environment variables
python-dotenv==1.0.1

# MCP SDK
mcp>=1.1.1

# LLM API
anthropic>=0.40.0

# Async support
aiohttp==3.9.1
anyio>=4.0.0

# Utilities
python-json-logger==2.0.7
pydantic>=2.5.2